+++ src/components/BookList.tsx
@@ -18,6 +18,7 @@ import * as DS from "@nypl/design-system-react-components";
 import MediumIndicator from "components/MediumIndicator";
 import { ArrowForward } from "icons";
 import useIsBorrowed from "hooks/useIsBorrowed";
+import BookCover from "./BookCover";
 
 /**
  * In a collection you can:
@@ -73,7 +74,17 @@ export const BookListItem: React.FC<{ book: BookData }> = ({ book }) => {
     >
       <DS.Card
         sx={{ bg: "ui.white" }}
-        image={<DS.Image src={book.imageUrl ?? ""} isDecorative />}
+        image={
+          <BookCover
+            book={book}
+            sx={{
+              height: "100%",
+              width: "100%",
+              maxHeight: "100%",
+              maxWidth: "100%"
+            }}
+          />
+        }
         ctas={
           <div
             sx={{